+++ code/libs/electron/src/bin/common/Settings.ts
@@ -1,5 +1,5 @@
-import { IUIHarnessElectronConfig, IUIHarnessEntry } from '../types';
-import { fs, fsPath, log, jsYaml, value, Package, config } from './common';
+import { IUIHarnessElectronConfig, IUIHarnessEntry } from '../../types';
+import { fs, fsPath, log, jsYaml, value, Package, config } from './libs';
 
 export { Package };
 const UIHARNESS_YAML = 'uiharness.yml';
