+++ code/libs/electron/src/bin/common/util.ts
@@ -1,18 +1,5 @@
-import { Package, ParcelBundler, fsPath, log } from './common';
-import { Settings } from './Settings';
-
-/**
- * Creates a new Parcel bundler.
- */
-export function createParcelBundler(settings: Settings) {
-  const entryFiles = settings.entries.map(e => e.html.absolute);
-  const args = settings.buildArgs;
-  return new ParcelBundler(entryFiles, {
-    sourceMaps: args.sourcemaps,
-    scopeHoist: args.treeshake,
-    target: 'browser',
-  });
-}
+import { log, fsPath } from './libs';
+import { Settings, Package } from './Settings';
 
 /**
  * Logs common information about the module.
@@ -21,8 +8,9 @@ export function logInfo(args: {
   settings: Settings;
   pkg: Package;
   port?: boolean | number;
+  main?: string;
 }) {
-  const { settings, pkg } = args;
+  const { settings, pkg, main } = args;
   const ROOT_DIR = fsPath.resolve('.');
 
   const formatPath = (path: string) => {
@@ -32,17 +20,14 @@ export function logInfo(args: {
     return `${dir}/${log.cyan(file)}`;
   };
 
-  const entryFiles = settings.entries.map(e => e.html.absolute);
+  // const entryFiles = settings.entries.map(e => e.html.absolute);
   const showPort = Boolean(args.port);
   const port = typeof args.port === 'number' ? args.port : settings.port;
 
   log.info();
   log.info.gray(`package: ${log.magenta(pkg.name)}`);
   log.info.gray(`version: ${pkg.version}`);
-  log.info.gray(`entry:   ${formatPath(entryFiles[0])}`);
-  entryFiles.slice(1).forEach(path => {
-    log.info.gray(`         ${formatPath(path)}`);
-  });
+  log.info.gray(`main:    ${formatPath(main || '')}`);
   if (showPort) {
     log.info.gray(`port:    ${log.yellow(port)}`);
   }
