+++ code/libs/web/src/bin/common/libs.ts
@@ -1,8 +1,7 @@
+import * as filesize from 'filesize';
 import * as ParcelBundler from 'parcel-bundler';
-import * as yargs from 'yargs';
-import * as execa from 'execa';
 
-export { ParcelBundler, yargs, execa };
+export { filesize, ParcelBundler };
 export {
   config,
   fs,
@@ -12,4 +11,4 @@ export {
   jsYaml,
   Package,
 } from '@uiharness/core/lib/server';
-export * from '../common';
+export * from '../../common';
