+++ code/libs/web/src/bin/common/util.ts
@@ -1,5 +1,18 @@
-import { log, fsPath } from './common';
-import { Settings, Package } from './Settings';
+import { Package, ParcelBundler, fsPath, log } from './libs';
+import { Settings } from './Settings';
+
+/**
+ * Creates a new Parcel bundler.
+ */
+export function createParcelBundler(settings: Settings) {
+  const entryFiles = settings.entries.map(e => e.html.absolute);
+  const args = settings.buildArgs;
+  return new ParcelBundler(entryFiles, {
+    sourceMaps: args.sourcemaps,
+    scopeHoist: args.treeshake,
+    target: 'browser',
+  });
+}
 
 /**
  * Logs common information about the module.
@@ -8,9 +21,8 @@ export function logInfo(args: {
   settings: Settings;
   pkg: Package;
   port?: boolean | number;
-  main?: string;
 }) {
-  const { settings, pkg, main } = args;
+  const { settings, pkg } = args;
   const ROOT_DIR = fsPath.resolve('.');
 
   const formatPath = (path: string) => {
@@ -20,14 +32,17 @@ export function logInfo(args: {
     return `${dir}/${log.cyan(file)}`;
   };
 
-  // const entryFiles = settings.entries.map(e => e.html.absolute);
+  const entryFiles = settings.entries.map(e => e.html.absolute);
   const showPort = Boolean(args.port);
   const port = typeof args.port === 'number' ? args.port : settings.port;
 
   log.info();
   log.info.gray(`package: ${log.magenta(pkg.name)}`);
   log.info.gray(`version: ${pkg.version}`);
-  log.info.gray(`main:    ${formatPath(main || '')}`);
+  log.info.gray(`entry:   ${formatPath(entryFiles[0])}`);
+  entryFiles.slice(1).forEach(path => {
+    log.info.gray(`         ${formatPath(path)}`);
+  });
   if (showPort) {
     log.info.gray(`port:    ${log.yellow(port)}`);
   }
